pdf_parser/requirements.txt
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
pdfminer.six==20220524
pycparser==2.21
beautifulsoup4==4.7.1
